Day 23
The leaderboard free fall continues. 😅
For part 1 I read the puzzle too quickly and though we were looking for
trees of size 3 in a forest, which I solved with union find.
After this failed on the sample, I quickly managed to write a correct
algorithm for part 1.
For part 2 I looked up "network maxclique" on DDG, and unfortunately
only found the approximation algorithm, which gives the wrong result.
This coincided with the part 2 leaderboard filling up.