Day 24
3 points ⭐
I wasted a lot of time on z3 for part 2.
When I switched to an incremental bit-by-bit randomized approach, I used
way too many random trials (I think I used 1<<20 or something, for the
solve I used 1<<12), which caused me to spend a lot of time on
optimizing other parts of the approach.